backend/requirements.txt
fastapi
uvicorn
motor
pymongo
python-dotenv
bcrypt
PyJWT
pydantic
pydantic-settings
python-multipart
